requirements.txt
pytest>=7.0
python-dotenv>=1.0

# UI
selenium>=4.0
webdriver-manager>=4.0

# API
requests>=2.28.0